requirements.txt
# Manipulação de Dados e Arrays
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Processamento Geoespacial (Crítico para definir os Nós/Bairros)
geopandas>=0.13.0
shapely>=2.0.0
rtree>=1.0.1
osmnx>=1.9.0  # Para baixar geometrias de bairros via OpenStreetMap
folium>=0.14.0 # Para visualização rápida de mapas nos notebooks

# Deep Learning (Graph Neural Networks)
# Nota: Instalar versão compatível com seu CUDA. Exemplo base CPU/CUDA 11.8:
torch>=2.0.0
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0

# Machine Learning Geral
scikit-learn>=1.2.0

# Web Framework e Interface
streamlit>=1.28.0
streamlit-folium>=0.11.0

# IA e LLM (Google Gemini)
google-generativeai>=0.3.0
google-api-core>=2.11.0

# Geração de PDFs
reportlab>=4.0.0

# Utilitários e Sistema
python-dotenv>=1.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
networkx>=3.1  # Manipulação de grafos clássica
requests>=2.31.0